Detach Authentication functions into AuthenticationMixin
